pyproject.toml
[project]
name = "rag-research-nextjs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.2.1",
    "beautifulsoup4>=4.12.3",
    "commonmark>=0.9.1",
    "dspy==2.5.43",
    "fastapi[standard]>=0.111.1",
    "lancedb>=0.17.0",
    "markdown-it-py>=3.0.0",
    "markdownify>=0.14.1",
    "modal>=0.71.2",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sentence-transformers>=3.3.1",
    "tantivy>=0.22.0",
    "torch>=2.5.1",
    "transformers",
    "uvicorn>=0.22.0",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git" }
